Bugfix to _make_move() - swap move changes board colours, not players
Previously, the `_make_move()` would swap the players:
https://github.com/mingfeisun/COMP34111-AI-Games-Hex/blob/main/src/Game.py#L184C5-L202C71

However, there is also a call to swap the players in `_play()`:
https://github.com/mingfeisun/COMP34111-AI-Games-Hex/blob/caadccb96c08b98776f6856f698f5e004850614d/src/Game.py#L181

As a result, the players would get into a muddle, if either player makes a swap move. Specifically, the vertical player would need to start playing horizontally, and the horizontal player would need to start playing vertically.

It is more coherent to edit the board instead of the players. 

There are already assertions in place, in `_play()`, to prevent modifications to the board by the player, which occur after a move is generated, but before the move is actioned. These assertions are then reset at the start of each turn. As a result, no further edits are required to `Game.py`, besides the one in `_make_move()` .